+++ src/main/kotlin/com/github/bobi/osgiannotationprocessor/settings/OSGIScrProjectSettings.kt
@@ -1,8 +1,6 @@
 package com.github.bobi.osgiannotationprocessor.settings
 
-import com.intellij.openapi.components.PersistentStateComponent
-import com.intellij.openapi.components.State
-import com.intellij.openapi.components.Storage
+import com.intellij.openapi.components.*
 import com.intellij.openapi.project.Project
 import com.intellij.util.xmlb.annotations.OptionTag
 
@@ -15,6 +13,7 @@ import com.intellij.util.xmlb.annotations.OptionTag
     name = "OsgiScrSettings",
     storages = [(Storage("osgi-scr.xml"))]
 )
+@Service(Service.Level.PROJECT)
 class OSGIScrProjectSettings : PersistentStateComponent<OSGIScrProjectSettings>, OSGIScrConfiguration {
 
     @OptionTag
@@ -61,9 +60,6 @@ class OSGIScrProjectSettings : PersistentStateComponent<OSGIScrProjectSettings>,
     }
 
     companion object {
-
-        fun getInstance(project: Project): OSGIScrProjectSettings =
-            project.getService(OSGIScrProjectSettings::class.java)
-
+        fun getInstance(project: Project): OSGIScrProjectSettings = project.service()
     }
 }